requirements.txt
ultralytics==8.3.175
torch>=1.7
Pillow
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles>=23.2.1